refactor: use enum for order status
