+++ src/adapter/driven/infra/sqa_models/order_persistent_model.py
@@ -1,10 +1,11 @@
-from sqlalchemy import Column, ForeignKey, String
-from sqlalchemy.orm import relationship
+from sqlalchemy import Column, ForeignKey
+from sqlalchemy import Enum as SaEnum
+from sqlalchemy.orm import Mapped, relationship
 
-from src.core.domain.entities.order import Order as OrderEntity
-from src.core.domain.value_objects.order_status import OrderStatus
+from src.core.domain.entities import Order as OrderEntity
+from src.core.domain.value_objects import OrderStatus
 
-from ..sqa_models.persistent_model import PersistentModel
+from .persistent_model import PersistentModel
 
 
 class OrderPersistentModel(PersistentModel):
@@ -13,7 +14,7 @@ class OrderPersistentModel(PersistentModel):
     __tablename__ = "orders"
 
     user_uuid = Column(ForeignKey("customers.uuid"), nullable=False)
-    status = Column(String, nullable=False)
+    status: Mapped[OrderStatus] = Column(SaEnum(OrderStatus), nullable=False)
     products = relationship("OrderProductPersistentModel", back_populates="order")
 
     def to_entity(self) -> OrderEntity:
@@ -21,7 +22,7 @@ class OrderPersistentModel(PersistentModel):
         return OrderEntity(
             user_uuid=self.user_uuid,
             products=[product.to_entity() for product in self.products],
-            status=OrderStatus(self.status),
+            status=self.status,
             created_at=self.created_at,
             updated_at=self.updated_at,
             uuid=self.uuid,
@@ -32,8 +33,11 @@ class OrderPersistentModel(PersistentModel):
         """Converts an Order entity to the persistent model."""
         return OrderPersistentModel(
             user_uuid=entity.user_uuid,
-            status=entity.status.status,  # Converte o status para string
+            status=entity.status,
             uuid=entity.uuid,
             created_at=entity.created_at,
             updated_at=entity.updated_at,
         )
+
+
+__all__ = ["OrderPersistentModel"]
