+++ src/core/domain/entities/__init__.py
@@ -6,7 +6,8 @@ They form the core of the domain model, encapsulating both behavior and data.
 """
 
 from .customer import Customer
+from .order import Order
 from .order_product import OrderProduct
 from .product import Product
 
-__all__ = ["Customer", "OrderProduct", "Product"]
+__all__ = ["Customer", "Order", "OrderProduct", "Product"]
